requirements.txt
Flask==0.11.1
requests==2.5.3
lxml==3.4.2
cssselect==1.0.1
boto==2.43.0
Pillow==4.0.0
line-bot-sdk==1.1.0
gunicorn==19.6.0
